chore(indexer): split large rust modules
