+++ src/datalens/mod.rs
@@ -1,0 +1,15 @@
+mod client;
+mod query;
+mod retry;
+mod types;
+
+pub use client::DatalensHttpClient;
+pub use query::{
+    evm_chain_name, logs_from_native_query_payload, native_graphql_request,
+    native_graphql_transaction_request, transactions_from_native_query_payload, tron_chain_name,
+};
+pub use retry::{DatalensFailureKind, classify_datalens_failure_message};
+pub use types::{
+    DatalensLog, DatalensLogQuery, DatalensLogQueryResult, DatalensLogReader, DatalensTransaction,
+    DatalensTransactionQuery, DatalensTransactionQueryResult,
+};
