+++ src/schema/mod.rs
@@ -1,0 +1,21 @@
+mod assignments;
+mod rows;
+mod tables;
+mod types;
+
+pub use assignments::{
+    ADDRESS_ORACLE, ADDRESS_RELAYER, AssignmentConfig, AssignmentUpdate,
+    LEGACY_B49E_ARBITRUM_FROM_BLOCK, LEGACY_B49E_DARWINIA_FROM_BLOCK, LEGACY_B49E_ORACLE,
+    LEGACY_B49E_ORACLE_FROM_BLOCK, LEGACY_MIXED_CASE_ACCEPTED_ID,
+    LEGACY_MIXED_CASE_ACCEPTED_ORACLE, accepted_oracle_value, apply_assignment_to_accepted,
+    is_oracle_assignment_for_accepted,
+};
+pub use rows::{
+    MsgportMessageRecvRow, MsgportMessageSentRow, OrmpHashImportedRow, OrmpMessageAcceptedRow,
+    OrmpMessageAssignedRow, OrmpMessageDispatchedRow, SignaturePubSignatureSubmittionRow,
+};
+pub use tables::{LegacySchema, POSTGRES_SCHEMA_MIGRATION};
+pub use types::{
+    ChainLogMetadata, EventSource, LegacyColumn, LegacyColumnType, LegacyIdRule, LegacyOrmPEvent,
+    LegacyTable,
+};
