+++ tests/legacy_parity_support/mod.rs
@@ -1,0 +1,9 @@
+mod fixtures;
+mod rows;
+mod values;
+
+pub use fixtures::{evm_fixture_logs, tron_fixture_logs};
+pub use rows::{
+    compatibility_rows, evm_expected_rows, fetch_compatibility_rows, legacy_expected_rows,
+    test_database_url, tron_expected_rows, truncate_legacy_tables,
+};
