+++ tests/runtime_runner.rs
@@ -1,324 +1,25 @@
 use std::{
     collections::BTreeMap,
-    sync::{Arc, Mutex},
     time::{Duration, Instant},
 };
 
 use ormpindexer::{
-    checkpoint::{Checkpoint, CheckpointStore, InMemoryCheckpointStore, plan_next_range},
+    checkpoint::{CheckpointStore, InMemoryCheckpointStore},
     config::{FinalityMode, RuntimeConfig},
-    database::{DryRunEventWriter, EventWriter},
-    datalens::{
-        DatalensLog, DatalensLogQuery, DatalensLogQueryResult, DatalensLogReader,
-        DatalensTransaction, DatalensTransactionQuery, DatalensTransactionQueryResult,
-    },
-    decoder::{EventDecoder, NoopDecoder},
+    database::DryRunEventWriter,
+    datalens::{DatalensLog, DatalensTransaction, DatalensTransactionQuery},
+    decoder::NoopDecoder,
     planner::MSGPORT_MESSAGE_SENT_TOPIC,
     runner::{IndexerRunner, RunnerReport},
-    schema::{ChainLogMetadata, EventSource, LegacyOrmPEvent},
+    schema::LegacyOrmPEvent,
 };
 
-#[test]
-fn test_runtime_config_from_env_map_reads_datalens_database_and_chain_settings() {
-    let env = BTreeMap::from([
-        (
-            "ORMPINDEXER_DATALENS_ENDPOINT".to_owned(),
-            "https://datalens.example".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_APPLICATION".to_owned(),
-            "ormp-production".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_TOKEN".to_owned(),
-            "secret-token".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_TIMEOUT_SECS".to_owned(),
-            "120".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_QUERY_MAX_ATTEMPTS".to_owned(),
-            "5".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_HEAD_BUFFER_BLOCKS".to_owned(),
-            "2".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_MIN_REQUEST_INTERVAL_MS".to_owned(),
-            "250".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATABASE_URL".to_owned(),
-            "postgres://user:pass@localhost/ormp".to_owned(),
-        ),
-        ("ORMPINDEXER_ENABLED_CHAINS".to_owned(), "1,46".to_owned()),
-        ("ORMPINDEXER_BATCH_SIZE".to_owned(), "250".to_owned()),
-        ("ORMPINDEXER_START_BLOCK".to_owned(), "1000".to_owned()),
-        ("ORMPINDEXER_FINALITY_MODE".to_owned(), "durable".to_owned()),
-        (
-            "ORMPINDEXER_DATALENS_WARMUP_ENABLED".to_owned(),
-            "true".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_WARMUP_ENSURE_ON_STARTUP".to_owned(),
-            "false".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_WARMUP_REQUIRED".to_owned(),
-            "true".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_WARMUP_CHUNK_SIZE".to_owned(),
-            "500".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_WARMUP_END_BLOCK".to_owned(),
-            "9000".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_CHAIN_1_CONTRACTS".to_owned(),
-            "0x111,0x222".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_CHAIN_1_TOPICS".to_owned(),
-            "0xaaa,0xbbb".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_CHAIN_46_START_BLOCK".to_owned(),
-            "2000".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_CHAIN_46_BATCH_SIZE".to_owned(),
-            "50".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_CHAIN_46_CONTRACTS".to_owned(),
-            "0x333".to_owned(),
-        ),
-    ]);
+mod runtime_support;
 
-    let config = RuntimeConfig::from_env_map(&env).expect("config parses");
-
-    assert_eq!(config.datalens.endpoint, "https://datalens.example");
-    assert_eq!(config.datalens.application, "ormp-production");
-    assert!(config.datalens.token.is_some());
-    assert_eq!(config.datalens.timeout, Duration::from_secs(120));
-    assert_eq!(config.datalens.query_max_attempts, 5);
-    assert_eq!(config.datalens.head_buffer_blocks, 2);
-    assert_eq!(
-        config.datalens.min_request_interval,
-        Duration::from_millis(250)
-    );
-    assert!(config.database_url.is_some());
-    assert_eq!(config.enabled_chains.len(), 2);
-    assert_eq!(config.batch_size, 250);
-    assert_eq!(config.start_block, 1000);
-    assert_eq!(config.finality_mode, FinalityMode::Durable);
-    assert!(config.warmup.enabled);
-    assert!(!config.warmup.ensure_on_startup);
-    assert!(config.warmup.required);
-    assert_eq!(config.warmup.chunk_size, 500);
-    assert_eq!(config.warmup.end_block, Some(9000));
-    assert_eq!(
-        config.chain(1).expect("chain 1").contracts,
-        vec!["0x111", "0x222"]
-    );
-    assert_eq!(
-        config.chain(1).expect("chain 1").topics,
-        vec!["0xaaa", "0xbbb"]
-    );
-    assert_eq!(config.chain(46).expect("chain 46").start_block, 2000);
-    assert_eq!(config.chain(1).expect("chain 1").batch_size, 250);
-    assert_eq!(config.chain(46).expect("chain 46").batch_size, 50);
-}
-
-#[test]
-fn test_runtime_config_from_env_map_reads_warmup_defaults() {
-    let env = BTreeMap::from([
-        (
-            "ORMPINDEXER_DATALENS_ENDPOINT".to_owned(),
-            "https://datalens.example".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_APPLICATION".to_owned(),
-            "ormp-production".to_owned(),
-        ),
-        ("ORMPINDEXER_ENABLED_CHAINS".to_owned(), "46".to_owned()),
-        ("ORMPINDEXER_BATCH_SIZE".to_owned(), "250".to_owned()),
-    ]);
-
-    let config = RuntimeConfig::from_env_map(&env).expect("config parses");
-
-    assert!(config.warmup.enabled);
-    assert!(config.warmup.ensure_on_startup);
-    assert!(!config.warmup.required);
-    assert_eq!(config.warmup.chunk_size, 250);
-    assert_eq!(config.warmup.end_block, None);
-}
-
-#[test]
-fn test_runtime_config_reads_datalens_retry_defaults() {
-    let env = BTreeMap::from([
-        (
-            "ORMPINDEXER_DATALENS_ENDPOINT".to_owned(),
-            "https://datalens.example".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_APPLICATION".to_owned(),
-            "ormp-production".to_owned(),
-        ),
-        ("ORMPINDEXER_ENABLED_CHAINS".to_owned(), "46".to_owned()),
-    ]);
-
-    let config = RuntimeConfig::from_env_map(&env).expect("config parses");
-
-    assert_eq!(config.datalens.timeout, Duration::from_secs(300));
-    assert_eq!(config.datalens.query_max_attempts, 3);
-    assert_eq!(config.datalens.head_buffer_blocks, 1);
-    assert_eq!(
-        config.datalens.min_request_interval,
-        Duration::from_millis(100)
-    );
-    assert_eq!(config.batch_size, 1_000);
-    assert_eq!(config.chain(46).expect("chain 46").batch_size, 1_000);
-}
-
-#[test]
-fn test_runtime_config_rejects_zero_chain_batch_size() {
-    let env = BTreeMap::from([
-        (
-            "ORMPINDEXER_DATALENS_ENDPOINT".to_owned(),
-            "https://datalens.example".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_APPLICATION".to_owned(),
-            "ormp-production".to_owned(),
-        ),
-        ("ORMPINDEXER_ENABLED_CHAINS".to_owned(), "46".to_owned()),
-        ("ORMPINDEXER_CHAIN_46_BATCH_SIZE".to_owned(), "0".to_owned()),
-    ]);
-
-    let error = RuntimeConfig::from_env_map(&env).expect_err("zero chain batch size is invalid");
-
-    assert!(
-        error
-            .to_string()
-            .contains("ORMPINDEXER_CHAIN_46_BATCH_SIZE must be greater than zero")
-    );
-}
-
-#[test]
-fn test_runtime_config_rejects_zero_datalens_timeout() {
-    let env = BTreeMap::from([
-        (
-            "ORMPINDEXER_DATALENS_ENDPOINT".to_owned(),
-            "https://datalens.example".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_APPLICATION".to_owned(),
-            "ormp-production".to_owned(),
-        ),
-        ("ORMPINDEXER_ENABLED_CHAINS".to_owned(), "46".to_owned()),
-        (
-            "ORMPINDEXER_DATALENS_TIMEOUT_SECS".to_owned(),
-            "0".to_owned(),
-        ),
-    ]);
-
-    let error = RuntimeConfig::from_env_map(&env).expect_err("zero timeout is invalid");
-
-    assert!(
-        error
-            .to_string()
-            .contains("ORMPINDEXER_DATALENS_TIMEOUT_SECS must be greater than zero")
-    );
-}
-
-#[test]
-fn test_runtime_config_rejects_zero_datalens_query_max_attempts() {
-    let env = BTreeMap::from([
-        (
-            "ORMPINDEXER_DATALENS_ENDPOINT".to_owned(),
-            "https://datalens.example".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_APPLICATION".to_owned(),
-            "ormp-production".to_owned(),
-        ),
-        ("ORMPINDEXER_ENABLED_CHAINS".to_owned(), "46".to_owned()),
-        (
-            "ORMPINDEXER_DATALENS_QUERY_MAX_ATTEMPTS".to_owned(),
-            "0".to_owned(),
-        ),
-    ]);
-
-    let error = RuntimeConfig::from_env_map(&env).expect_err("zero attempts is invalid");
-
-    assert!(
-        error
-            .to_string()
-            .contains("ORMPINDEXER_DATALENS_QUERY_MAX_ATTEMPTS must be greater than zero")
-    );
-}
-
-#[test]
-fn test_runtime_config_rejects_zero_warmup_chunk_size() {
-    let env = BTreeMap::from([
-        (
-            "ORMPINDEXER_DATALENS_ENDPOINT".to_owned(),
-            "https://datalens.example".to_owned(),
-        ),
-        (
-            "ORMPINDEXER_DATALENS_APPLICATION".to_owned(),
-            "ormp-production".to_owned(),
-        ),
-        ("ORMPINDEXER_ENABLED_CHAINS".to_owned(), "46".to_owned()),
-        (
-            "ORMPINDEXER_DATALENS_WARMUP_CHUNK_SIZE".to_owned(),
-            "0".to_owned(),
-        ),
-    ]);
-
-    let error = RuntimeConfig::from_env_map(&env).expect_err("zero warmup chunk size is invalid");
-
-    assert!(
-        error
-            .to_string()
-            .contains("ORMPINDEXER_DATALENS_WARMUP_CHUNK_SIZE must be greater than zero")
-    );
-}
-
-#[test]
-fn test_database_url_from_env_map_does_not_require_runtime_chain_settings() {
-    let env = BTreeMap::from([(
-        "ORMPINDEXER_DATABASE_URL".to_owned(),
-        "postgres://user:pass@localhost/ormp".to_owned(),
-    )]);
-
-    let database_url =
-        RuntimeConfig::database_url_from_env_map(&env).expect("database URL parses independently");
-
-    assert_eq!(
-        database_url.expose_secret(),
-        "postgres://user:pass@localhost/ormp"
-    );
-}
-
-#[test]
-fn test_plan_next_range_uses_checkpoint_and_batch_size() {
-    let checkpoint = Checkpoint {
-        chain_id: 46,
-        dataset: "datalens-native".to_owned(),
-        next_block: 120,
-    };
-
-    let range = plan_next_range(&checkpoint, 25).expect("range is planned");
-
-    assert_eq!(range.from_block, 120);
-    assert_eq!(range.to_block, 144);
-}
+use runtime_support::{
+    EchoTransactionFromDecoder, FailingEventWriter, FailingEventWriterWithMessage,
+    RecordingDatalensReader, RecordingEventWriter,
+};
 
 #[tokio::test]
 async fn test_runner_successful_batch_advances_checkpoint_to_next_range() {
@@ -1065,185 +766,4 @@ async fn test_runner_loop_recovers_chain_errors_without_blocking_other_chains() 
 
     assert_eq!(checkpoints.next_block(1, "evm.logs").await.unwrap(), 10);
     run.abort();
-}
-
-#[derive(Clone)]
-struct RecordingDatalensReader {
-    logs: Vec<DatalensLog>,
-    transactions: Vec<DatalensTransaction>,
-    heads: BTreeMap<u64, u64>,
-    query_delays: BTreeMap<u64, Duration>,
-    query_failures: BTreeMap<u64, String>,
-    range_query_failures: BTreeMap<(u64, u64, u64), String>,
-    queries: Arc<Mutex<Vec<DatalensLogQuery>>>,
-    transaction_queries: Arc<Mutex<Vec<DatalensTransactionQuery>>>,
-}
-
-impl RecordingDatalensReader {
-    fn new(logs: Vec<DatalensLog>) -> Self {
-        Self {
-            logs,
-            transactions: Vec::new(),
-            heads: BTreeMap::new(),
-            query_delays: BTreeMap::new(),
-            query_failures: BTreeMap::new(),
-            range_query_failures: BTreeMap::new(),
-            queries: Arc::new(Mutex::new(Vec::new())),
-            transaction_queries: Arc::new(Mutex::new(Vec::new())),
-        }
-    }
-
-    fn with_head(mut self, chain_id: u64, head: u64) -> Self {
-        self.heads.insert(chain_id, head);
-        self
-    }
-
-    fn with_transactions(mut self, transactions: Vec<DatalensTransaction>) -> Self {
-        self.transactions = transactions;
-        self
-    }
-
-    fn transaction_queries(&self) -> Arc<Mutex<Vec<DatalensTransactionQuery>>> {
-        self.transaction_queries.clone()
-    }
-
-    fn with_query_delay(mut self, chain_id: u64, delay: Duration) -> Self {
-        self.query_delays.insert(chain_id, delay);
-        self
-    }
-
-    fn with_query_failure(mut self, chain_id: u64, error: &str) -> Self {
-        self.query_failures.insert(chain_id, error.to_owned());
-        self
-    }
-
-    fn with_range_query_failure(
-        mut self,
-        chain_id: u64,
-        from_block: u64,
-        to_block: u64,
-        error: &str,
-    ) -> Self {
-        self.range_query_failures
-            .insert((chain_id, from_block, to_block), error.to_owned());
-        self
-    }
-}
-
-impl DatalensLogReader for RecordingDatalensReader {
-    async fn latest_block(
-        &self,
-        chain_id: u64,
-        _finality_mode: FinalityMode,
-    ) -> anyhow::Result<u64> {
-        Ok(*self.heads.get(&chain_id).unwrap_or(&u64::MAX))
-    }
-
-    async fn query_logs(&self, query: DatalensLogQuery) -> anyhow::Result<DatalensLogQueryResult> {
-        if let Some(delay) = self.query_delays.get(&query.chain_id) {
-            tokio::time::sleep(*delay).await;
-        }
-        self.queries
-            .lock()
-            .expect("queries lock")
-            .push(query.clone());
-        if let Some(error) =
-            self.range_query_failures
-                .get(&(query.chain_id, query.from_block, query.to_block))
-        {
-            anyhow::bail!("{error}");
-        }
-        if let Some(error) = self.query_failures.get(&query.chain_id) {
-            anyhow::bail!("{error}");
-        }
-        Ok(DatalensLogQueryResult {
-            logs: self.logs.clone(),
-        })
-    }
-
-    async fn query_transactions(
-        &self,
-        query: DatalensTransactionQuery,
-    ) -> anyhow::Result<DatalensTransactionQueryResult> {
-        self.transaction_queries
-            .lock()
-            .expect("transaction queries lock")
-            .push(query.clone());
-        Ok(DatalensTransactionQueryResult {
-            transactions: self
-                .transactions
-                .iter()
-                .filter(|transaction| {
-                    transaction.block_number >= query.from_block
-                        && transaction.block_number <= query.to_block
-                })
-                .cloned()
-                .collect(),
-        })
-    }
-}
-
-struct EchoTransactionFromDecoder;
-
-impl EventDecoder for EchoTransactionFromDecoder {
-    async fn decode(&self, log: &DatalensLog) -> anyhow::Result<Vec<LegacyOrmPEvent>> {
-        Ok(vec![LegacyOrmPEvent::MsgportMessageSent {
-            metadata: ChainLogMetadata {
-                id: log.id.clone().expect("test log id"),
-                source: EventSource::Evm,
-                chain_id: log.chain_id.into(),
-                block_number: log.block_number.into(),
-                block_hash: log.block_hash.clone(),
-                block_timestamp: log.block_timestamp.expect("test timestamp").into(),
-                transaction_hash: log.transaction_hash.clone(),
-                transaction_index: log.transaction_index.expect("test transaction index"),
-                log_index: i32::try_from(log.log_index).expect("test log index"),
-                contract_address: log.address.clone(),
-                transaction_from: log.transaction_from.clone(),
-            },
-            msg_id: "0xmsgid".to_owned(),
-            from_dapp: "0xfromdapp".to_owned(),
-            to_chain_id: 1,
-            to_dapp: "0xtodapp".to_owned(),
-            message: "0xmessage".to_owned(),
-            params: "0xparams".to_owned(),
-        }])
-    }
-}
-
-#[derive(Clone, Default)]
-struct RecordingEventWriter {
-    events: Arc<Mutex<Vec<LegacyOrmPEvent>>>,
-}
-
-impl RecordingEventWriter {
-    fn events(&self) -> Vec<LegacyOrmPEvent> {
-        self.events.lock().expect("events lock").clone()
-    }
-}
-
-impl EventWriter for RecordingEventWriter {
-    async fn write_events(&self, events: &[LegacyOrmPEvent]) -> anyhow::Result<usize> {
-        self.events
-            .lock()
-            .expect("events lock")
-            .extend_from_slice(events);
-        Ok(events.len())
-    }
-}
-
-struct FailingEventWriter;
-
-impl EventWriter for FailingEventWriter {
-    async fn write_events(&self, _events: &[LegacyOrmPEvent]) -> anyhow::Result<usize> {
-        anyhow::bail!("write failed");
-    }
-}
-
-struct FailingEventWriterWithMessage(&'static str);
-
-impl EventWriter for FailingEventWriterWithMessage {
-    async fn write_events(&self, _events: &[LegacyOrmPEvent]) -> anyhow::Result<usize> {
-        anyhow::bail!(self.0);
-    }
 }
